Gutted the setup
Took out as much as possible to condense the Flask setup
