--- app/__init__.py
@@ -1,7 +1,0 @@
-from flask import Flask 
-from app.config import Config
-
-app = Flask(__name__)
-app.config.from_object(Config)
-
-from app import routes
