--- app/templates/data.html
@@ -1,10 +1,0 @@
-<form action="/data" method = "POST">
-    <p>Height in feet: <input type = "number" name = "BMI_params" value = "{{ feet }}"/></p>
-    <p>Height in inches: <input type = "number" name = "BMI_params" value = "{{ inches }}"/></p>
-    <p>Weight in pounds: <input type = "number" name = "BMI_params" value = "{{ pounds }}"/></p>
-    <p><input type = "submit" value = "Calculate" /></p>
-    {% if category %}
-    <p>At {{ feet }}' {{ inches }}'' and {{ pounds }} lbs, your BMI is {{ result }}.</p>
-    <p>Your category is {{ category }}.</p>
-    {% endif %}
-</form>
