--- app/templates/form.html
@@ -1,1 +1,0 @@
-<p>testing?</p>
